Add failFast Rename package Fix tests Fix style
