+++ src/main/java/org/prebid/server/spring/config/bidder/EmxDigitalConfiguration.java
@@ -1,7 +1,7 @@
 package org.prebid.server.spring.config.bidder;
 
 import org.prebid.server.bidder.BidderDeps;
-import org.prebid.server.bidder.emx_digital.EmxDigitalBidder;
+import org.prebid.server.bidder.emxdigital.EmxDigitalBidder;
 import org.prebid.server.spring.config.bidder.model.BidderConfigurationProperties;
 import org.prebid.server.spring.config.bidder.util.BidderDepsAssembler;
 import org.prebid.server.spring.config.bidder.util.BidderInfoCreator;
